removes a bunch of unnecessary tests
