+++ recipe/emailpassword/loginmethods_test.go
@@ -1,4 +1,4 @@
-package thirdpartyemailpassword
+package emailpassword
 
 import (
 	"encoding/json"
