--- recipe/thirdpartyemailpassword/authorisationUrlFeature_test.go
@@ -1,164 +1,0 @@
-/* Copyright (c) 2021, VRAI Labs and/or its affiliates. All rights reserved.
- *
- * This software is licensed under the Apache License, Version 2.0 (the
- * "License") as published by the Apache Software Foundation.
- *
- * You may not use this file except in compliance with the License. You may
- * obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- */
-
-package thirdpartyemailpassword
-
-import (
-	"encoding/json"
-	"io"
-	"net/http"
-	"net/http/httptest"
-	"net/url"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/supertokens/supertokens-golang/recipe/session"
-	"github.com/supertokens/supertokens-golang/recipe/session/sessmodels"
-	"github.com/supertokens/supertokens-golang/recipe/thirdparty/tpmodels"
-	"github.com/supertokens/supertokens-golang/recipe/thirdpartyemailpassword/tpepmodels"
-	"github.com/supertokens/supertokens-golang/supertokens"
-	"github.com/supertokens/supertokens-golang/test/unittesting"
-)
-
-func TestMinimumConfigForThirdPartyModule(t *testing.T) {
-	customAntiCsrfVal := "VIA_TOKEN"
-	configValue := supertokens.TypeInput{
-		Supertokens: &supertokens.ConnectionInfo{
-			ConnectionURI: "http://localhost:8080",
-		},
-		AppInfo: supertokens.AppInfo{
-			APIDomain:     "api.supertokens.io",
-			AppName:       "SuperTokens",
-			WebsiteDomain: "supertokens.io",
-		},
-		RecipeList: []supertokens.Recipe{
-			Init(
-				&tpepmodels.TypeInput{
-					Providers: []tpmodels.ProviderInput{
-						customProvider1,
-					},
-				},
-			),
-			session.Init(&sessmodels.TypeInput{
-				AntiCsrf: &customAntiCsrfVal,
-				GetTokenTransferMethod: func(req *http.Request, forCreateNewSession bool, userContext supertokens.UserContext) sessmodels.TokenTransferMethod {
-					return sessmodels.CookieTransferMethod
-				},
-			}),
-		},
-	}
-
-	BeforeEach()
-	unittesting.StartUpST("localhost", "8080")
-	defer AfterEach()
-	err := supertokens.Init(configValue)
-
-	if err != nil {
-		t.Error(err.Error())
-	}
-
-	mux := http.NewServeMux()
-	testServer := httptest.NewServer(supertokens.Middleware(mux))
-	defer testServer.Close()
-
-	resp, err := http.Get(testServer.URL + "/auth/authorisationurl?thirdPartyId=custom")
-	if err != nil {
-		t.Error(err.Error())
-	}
-
-	dataInBytes, err := io.ReadAll(resp.Body)
-	if err != nil {
-		t.Error(err.Error())
-	}
-	resp.Body.Close()
-
-	var data map[string]interface{}
-	err = json.Unmarshal(dataInBytes, &data)
-	if err != nil {
-		t.Error(err.Error())
-	}
-
-	assert.Equal(t, "OK", data["status"])
-
-	fetchedUrl, err := url.Parse(data["urlWithQueryParams"].(string))
-	if err != nil {
-		t.Error(err.Error())
-	}
-	assert.Equal(t, "test.com", fetchedUrl.Host)
-	assert.Equal(t, "/oauth/auth", fetchedUrl.Path)
-}
-
-func TestThirdPartyProviderDoesnotExsit(t *testing.T) {
-	customAntiCsrfVal := "VIA_TOKEN"
-	configValue := supertokens.TypeInput{
-		Supertokens: &supertokens.ConnectionInfo{
-			ConnectionURI: "http://localhost:8080",
-		},
-		AppInfo: supertokens.AppInfo{
-			APIDomain:     "api.supertokens.io",
-			AppName:       "SuperTokens",
-			WebsiteDomain: "supertokens.io",
-		},
-		RecipeList: []supertokens.Recipe{
-			Init(
-				&tpepmodels.TypeInput{
-					Providers: []tpmodels.ProviderInput{
-						customProvider1,
-					},
-				},
-			),
-			session.Init(&sessmodels.TypeInput{
-				AntiCsrf: &customAntiCsrfVal,
-				GetTokenTransferMethod: func(req *http.Request, forCreateNewSession bool, userContext supertokens.UserContext) sessmodels.TokenTransferMethod {
-					return sessmodels.CookieTransferMethod
-				},
-			}),
-		},
-	}
-
-	BeforeEach()
-	unittesting.StartUpST("localhost", "8080")
-	defer AfterEach()
-	err := supertokens.Init(configValue)
-
-	if err != nil {
-		t.Error(err.Error())
-	}
-
-	mux := http.NewServeMux()
-	testServer := httptest.NewServer(supertokens.Middleware(mux))
-	defer testServer.Close()
-
-	resp, err := http.Get(testServer.URL + "/auth/authorisationurl?thirdPartyId=google")
-	if err != nil {
-		t.Error(err.Error())
-	}
-
-	assert.Equal(t, http.StatusBadRequest, resp.StatusCode)
-
-	dataInBytes, err := io.ReadAll(resp.Body)
-	if err != nil {
-		t.Error(err.Error())
-	}
-	resp.Body.Close()
-
-	var data map[string]interface{}
-	err = json.Unmarshal(dataInBytes, &data)
-	if err != nil {
-		t.Error(err.Error())
-	}
-
-	assert.Equal(t, "the provider google could not be found in the configuration", data["message"].(string))
-}
