--- recipe/thirdpartyemailpassword/emailExist_test.go
@@ -1,222 +1,0 @@
-/* Copyright (c) 2021, VRAI Labs and/or its affiliates. All rights reserved.
- *
- * This software is licensed under the Apache License, Version 2.0 (the
- * "License") as published by the Apache Software Foundation.
- *
- * You may not use this file except in compliance with the License. You may
- * obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- */
-
-package thirdpartyemailpassword
-
-import (
-	"net/http"
-	"net/http/httptest"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/supertokens/supertokens-golang/recipe/session"
-	"github.com/supertokens/supertokens-golang/recipe/session/sessmodels"
-	"github.com/supertokens/supertokens-golang/recipe/thirdpartyemailpassword/tpepmodels"
-	"github.com/supertokens/supertokens-golang/supertokens"
-	"github.com/supertokens/supertokens-golang/test/unittesting"
-)
-
-func TestIfDisableAPIdefaultEmailExistDoesNotWork(t *testing.T) {
-	configValue := supertokens.TypeInput{
-		Supertokens: &supertokens.ConnectionInfo{
-			ConnectionURI: "http://localhost:8080",
-		},
-		AppInfo: supertokens.AppInfo{
-			APIDomain:     "api.supertokens.io",
-			AppName:       "SuperTokens",
-			WebsiteDomain: "supertokens.io",
-		},
-		RecipeList: []supertokens.Recipe{
-			Init(&tpepmodels.TypeInput{
-				Override: &tpepmodels.OverrideStruct{
-					APIs: func(originalImplementation tpepmodels.APIInterface) tpepmodels.APIInterface {
-						*originalImplementation.EmailPasswordEmailExistsGET = nil
-						return originalImplementation
-					},
-				},
-			}),
-		},
-	}
-
-	BeforeEach()
-	unittesting.StartUpST("localhost", "8080")
-	defer AfterEach()
-	err := supertokens.Init(configValue)
-	if err != nil {
-		t.Error(err.Error())
-	}
-	mux := http.NewServeMux()
-	testServer := httptest.NewServer(supertokens.Middleware(mux))
-	defer testServer.Close()
-
-	req, err := http.NewRequest(http.MethodGet, testServer.URL+"/auth/signup/email/exists", nil)
-	q := req.URL.Query()
-	q.Add("email", "random@gmail.com")
-	req.URL.RawQuery = q.Encode()
-	assert.NoError(t, err)
-	res, err := http.DefaultClient.Do(req)
-	assert.NoError(t, err)
-	assert.Equal(t, http.StatusNotFound, res.StatusCode)
-}
-
-func TestGoodInputEmailExist(t *testing.T) {
-	configValue := supertokens.TypeInput{
-		Supertokens: &supertokens.ConnectionInfo{
-			ConnectionURI: "http://localhost:8080",
-		},
-		AppInfo: supertokens.AppInfo{
-			APIDomain:     "api.supertokens.io",
-			AppName:       "SuperTokens",
-			WebsiteDomain: "supertokens.io",
-		},
-		RecipeList: []supertokens.Recipe{
-			Init(nil),
-			session.Init(&sessmodels.TypeInput{
-				GetTokenTransferMethod: func(req *http.Request, forCreateNewSession bool, userContext supertokens.UserContext) sessmodels.TokenTransferMethod {
-					return sessmodels.CookieTransferMethod
-				},
-			}),
-		},
-	}
-
-	BeforeEach()
-	unittesting.StartUpST("localhost", "8080")
-	defer AfterEach()
-	err := supertokens.Init(configValue)
-	if err != nil {
-		t.Error(err.Error())
-	}
-	mux := http.NewServeMux()
-	testServer := httptest.NewServer(supertokens.Middleware(mux))
-	defer testServer.Close()
-
-	resp, err := unittesting.SignupRequest("random@gmail.com", "validPass123", testServer.URL)
-	if err != nil {
-		t.Error(err.Error())
-	}
-
-	assert.Equal(t, http.StatusOK, resp.StatusCode)
-
-	result := *unittesting.HttpResponseToConsumableInformation(resp.Body)
-	assert.Equal(t, "OK", result["status"])
-
-	req, err := http.NewRequest(http.MethodGet, testServer.URL+"/auth/signup/email/exists", nil)
-	q := req.URL.Query()
-	q.Add("email", "random@gmail.com")
-	req.URL.RawQuery = q.Encode()
-	if err != nil {
-		t.Error(err.Error())
-	}
-	res, err := http.DefaultClient.Do(req)
-	if err != nil {
-		t.Error(err.Error())
-	}
-	result1 := *unittesting.HttpResponseToConsumableInformation(res.Body)
-	assert.Equal(t, http.StatusOK, res.StatusCode)
-	assert.Equal(t, "OK", result1["status"])
-	assert.Equal(t, true, result1["exists"])
-}
-
-func TestGoodInputEmailDoesNotExist(t *testing.T) {
-	configValue := supertokens.TypeInput{
-		Supertokens: &supertokens.ConnectionInfo{
-			ConnectionURI: "http://localhost:8080",
-		},
-		AppInfo: supertokens.AppInfo{
-			APIDomain:     "api.supertokens.io",
-			AppName:       "SuperTokens",
-			WebsiteDomain: "supertokens.io",
-		},
-		RecipeList: []supertokens.Recipe{
-			Init(nil),
-			session.Init(&sessmodels.TypeInput{
-				GetTokenTransferMethod: func(req *http.Request, forCreateNewSession bool, userContext supertokens.UserContext) sessmodels.TokenTransferMethod {
-					return sessmodels.CookieTransferMethod
-				},
-			}),
-		},
-	}
-
-	BeforeEach()
-	unittesting.StartUpST("localhost", "8080")
-	defer AfterEach()
-	err := supertokens.Init(configValue)
-	if err != nil {
-		t.Error(err.Error())
-	}
-	mux := http.NewServeMux()
-	testServer := httptest.NewServer(supertokens.Middleware(mux))
-	defer testServer.Close()
-
-	req, err := http.NewRequest(http.MethodGet, testServer.URL+"/auth/signup/email/exists", nil)
-	q := req.URL.Query()
-	q.Add("email", "random@gmail.com")
-	req.URL.RawQuery = q.Encode()
-	if err != nil {
-		t.Error(err.Error())
-	}
-	res, err := http.DefaultClient.Do(req)
-	if err != nil {
-		t.Error(err.Error())
-	}
-	result := *unittesting.HttpResponseToConsumableInformation(res.Body)
-	assert.Equal(t, http.StatusOK, res.StatusCode)
-	assert.Equal(t, "OK", result["status"])
-	assert.Equal(t, false, result["exists"])
-}
-
-func TestBadInputDoNotPassEmail(t *testing.T) {
-	configValue := supertokens.TypeInput{
-		Supertokens: &supertokens.ConnectionInfo{
-			ConnectionURI: "http://localhost:8080",
-		},
-		AppInfo: supertokens.AppInfo{
-			APIDomain:     "api.supertokens.io",
-			AppName:       "SuperTokens",
-			WebsiteDomain: "supertokens.io",
-		},
-		RecipeList: []supertokens.Recipe{
-			Init(nil),
-			session.Init(&sessmodels.TypeInput{
-				GetTokenTransferMethod: func(req *http.Request, forCreateNewSession bool, userContext supertokens.UserContext) sessmodels.TokenTransferMethod {
-					return sessmodels.CookieTransferMethod
-				},
-			}),
-		},
-	}
-
-	BeforeEach()
-	unittesting.StartUpST("localhost", "8080")
-	defer AfterEach()
-	err := supertokens.Init(configValue)
-	if err != nil {
-		t.Error(err.Error())
-	}
-	mux := http.NewServeMux()
-	testServer := httptest.NewServer(supertokens.Middleware(mux))
-	defer testServer.Close()
-
-	req, err := http.NewRequest(http.MethodGet, testServer.URL+"/auth/signup/email/exists", nil)
-	if err != nil {
-		t.Error(err.Error())
-	}
-	res, err := http.DefaultClient.Do(req)
-	if err != nil {
-		t.Error(err.Error())
-	}
-	result := *unittesting.HttpResponseToConsumableInformation(res.Body)
-	assert.Equal(t, http.StatusBadRequest, res.StatusCode)
-	assert.Equal(t, "Please provide the email as a GET param", result["message"])
-}
